Refactoring to make HttpClientFacade more reusable; add Currency Exchange Rate loading process, including in-memory cache

+++ ProfitWise.Data/Model/Currency.cs
@@ -2,7 +2,7 @@
 {
     public class Currency
     {
-        public long CurrencyId { get; set; }
+        public int CurrencyId { get; set; }
         public string Abbreviation { get; set; }
         public string Symbol { get; set; }
         public string Name { get; set; }
